WIP - Further work on the Guards on Employer Details
