+++ src/app/app.module.ts
@@ -15,6 +15,7 @@ import { MojHeaderComponent, MojHeaderNavigationItemComponent, GovukFooterCompon
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { httpErrorInterceptor } from '@interceptors';
 import { AppInitializerService } from '@services';
+import { FlowExitStateGuard } from './guards/flow-exit-state-guard/flow-exit-state.guard';
 
 @NgModule({
   declarations: [AppComponent],
@@ -41,6 +42,7 @@ import { AppInitializerService } from '@services';
       },
       deps: [AppInitializerService],
     },
+    [FlowExitStateGuard],
   ],
 
   bootstrap: [AppComponent],
