+++ src/app/constants/index.ts
@@ -1,12 +1,14 @@
 import { ACCOUNT_ENQUIRY_DEFAULT_STATE, ACCOUNT_ENQUIRY_DEFAULT_STATE_SEARCH } from './account-enquiry-default-state';
 import { API_PATHS } from './api-paths';
 import { MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE } from './manual-account-creation-employer-detail-state';
+import { MANUAL_ACCOUNT_CREATION_STATE } from './manual-account-creation-state';
 import { ROUTE_PERMISSIONS } from './route-permissions';
 
-export { 
-    ACCOUNT_ENQUIRY_DEFAULT_STATE, 
-    ACCOUNT_ENQUIRY_DEFAULT_STATE_SEARCH, 
-    ROUTE_PERMISSIONS, 
-    API_PATHS, 
-    MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE 
+export {
+  ACCOUNT_ENQUIRY_DEFAULT_STATE,
+  ACCOUNT_ENQUIRY_DEFAULT_STATE_SEARCH,
+  ROUTE_PERMISSIONS,
+  API_PATHS,
+  MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE,
+  MANUAL_ACCOUNT_CREATION_STATE,
 };
