+++ src/app/constants/manual-account-creation-employer-detail-state.ts
@@ -1,12 +1,12 @@
 export const MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE = {
-    employerName: null,
-    employeeReference: null,
-    employerEmailAddress: null,
-    employerTelephone: null,
-    employerAddress1: null,
-    employerAddress2: null,
-    employerAddress3: null,
-    employerAddress4: null,
-    employerAddress5: null,
-    employerPostcode: null
-}
+  employerName: null,
+  employeeReference: null,
+  employerEmailAddress: null,
+  employerTelephone: null,
+  employerAddress1: null,
+  employerAddress2: null,
+  employerAddress3: null,
+  employerAddress4: null,
+  employerAddress5: null,
+  employerPostcode: null,
+};
