+++ src/app/constants/manual-account-creation-state.ts
@@ -1,0 +1,5 @@
+import { MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE } from './manual-account-creation-employer-detail-state';
+
+export const MANUAL_ACCOUNT_CREATION_STATE = {
+  employerDetails: MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE,
+};
