+++ src/app/interfaces/manual-account-creation-employer-details-state.interface.ts
@@ -1,12 +1,12 @@
 export interface IManualAccountCreationEmployerDetailsState {
-    employerName: string | null;
-    employeeReference: string | null;
-    employerEmailAddress: string | null;
-    employerTelephone: string | null;
-    employerAddress1: string | null;
-    employerAddress2: string | null;
-    employerAddress3: string | null;
-    employerAddress4: string | null;
-    employerAddress5: string | null;
-    employerPostcode: string | null;
+  employerName: string | null;
+  employeeReference: string | null;
+  employerEmailAddress: string | null;
+  employerTelephone: string | null;
+  employerAddress1: string | null;
+  employerAddress2: string | null;
+  employerAddress3: string | null;
+  employerAddress4: string | null;
+  employerAddress5: string | null;
+  employerPostcode: string | null;
 }
