+++ src/app/interfaces/manual-account-creation-state.interface.ts
@@ -1,5 +1,5 @@
-import { IManualAccountCreationEmployerDetailsState } from "./manual-account-creation-employer-details-state.interface";
+import { IManualAccountCreationEmployerDetailsState } from './manual-account-creation-employer-details-state.interface';
 
-export interface IManualAccountCreation {
-    employerDetails: IManualAccountCreationEmployerDetailsState
+export interface IManualAccountCreationState {
+  employerDetails: IManualAccountCreationEmployerDetailsState;
 }
