+++ src/app/pages/manual-account-creation/create-account/create-account.component.ts
@@ -9,6 +9,7 @@ import {
   GovukTaskListItemComponent,
 } from '@components';
 import { ManualAccountCreationRoutes } from '@enums';
+import { StateService } from '@services';
 
 @Component({
   selector: 'app-create-account',
@@ -27,11 +28,13 @@ import { ManualAccountCreationRoutes } from '@enums';
 export class CreateAccountComponent {
   private readonly router = inject(Router);
   public readonly manualAccountCreationRoutes = ManualAccountCreationRoutes;
+  public readonly stateService = inject(StateService);
 
   /**
    * Handles back and navigates to the manual account creation page.
    */
   public handleBack(): void {
+    console.log(this.stateService.manualAccountCreation);
     this.router.navigate([ManualAccountCreationRoutes.accountDetails]);
   }
 }
