+++ src/app/services/state-service/state.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, signal } from '@angular/core';
-import { ACCOUNT_ENQUIRY_DEFAULT_STATE, MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE } from '@constants';
-import { IAccountEnquiryState, ILaunchDarklyConfig, IManualAccountCreationEmployerDetailsState, IUserState } from '@interfaces';
+import { ACCOUNT_ENQUIRY_DEFAULT_STATE, MANUAL_ACCOUNT_CREATION_STATE } from '@constants';
+import { IAccountEnquiryState, ILaunchDarklyConfig, IManualAccountCreationState, IUserState } from '@interfaces';
 import { LDFlagSet } from 'launchdarkly-js-client-sdk';
 
 @Injectable({
@@ -17,5 +17,5 @@ export class StateService {
   public ssoEnabled: boolean | null = false;
   public launchDarklyConfig: ILaunchDarklyConfig | null = null;
   public accountEnquiry: IAccountEnquiryState = ACCOUNT_ENQUIRY_DEFAULT_STATE;
-  public manualAccountCreationEmployerDetails: IManualAccountCreationEmployerDetailsState = MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE;
+  public manualAccountCreation: IManualAccountCreationState = MANUAL_ACCOUNT_CREATION_STATE;
 }
