+++ src/app/validators/index.ts
@@ -3,9 +3,9 @@ import { optionalEmailAddressValidator } from './optional-valid-email-address/op
 import { optionalPhoneNumberValidator } from './optional-valid-telephone/optional-valid-telephone.validator';
 import { overEighteenValidator } from './over-eighteen/over-eighteen.validator';
 
-export { 
-    overEighteenValidator,
-    optionalMaxLengthValidator,
-    optionalEmailAddressValidator, 
-    optionalPhoneNumberValidator,
+export {
+  overEighteenValidator,
+  optionalMaxLengthValidator,
+  optionalEmailAddressValidator,
+  optionalPhoneNumberValidator,
 };
