+++ src/app/validators/optional-max-length/optional-max-length.validator.ts
@@ -1,10 +1,10 @@
-import { ValidatorFn, AbstractControl, Validators } from "@angular/forms";
+import { ValidatorFn, AbstractControl, Validators } from '@angular/forms';
 
 export function optionalMaxLengthValidator(maxLength: number): ValidatorFn {
-    return (control: AbstractControl): { [key: string]: unknown } | null => {
-      if (control.value) {
-        return Validators.maxLength(maxLength)(control);
-      }
-      return null;
-    };
+  return (control: AbstractControl): { [key: string]: unknown } | null => {
+    if (control.value) {
+      return Validators.maxLength(maxLength)(control);
+    }
+    return null;
+  };
 }
