+++ src/app/validators/optional-valid-email-address/optional-valid-email-address.validator.ts
@@ -1,12 +1,12 @@
-import { ValidatorFn, AbstractControl } from "@angular/forms";
+import { ValidatorFn, AbstractControl } from '@angular/forms';
 
 export function optionalEmailAddressValidator(): ValidatorFn {
-    const emailPattern = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
-    return (control: AbstractControl): { [key: string]: unknown } | null => {
-        if (control.value) {
-            const valid = emailPattern.test(control.value);
-            return valid ? null : { 'emailPattern': { value: control.value } };
-        }
-        return null;
-    };
+  const emailPattern = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
+  return (control: AbstractControl): { [key: string]: unknown } | null => {
+    if (control.value) {
+      const valid = emailPattern.test(control.value);
+      return valid ? null : { emailPattern: { value: control.value } };
+    }
+    return null;
+  };
 }
